Added TextFieldView and changed EmailView and PasswordView
