+++ QuizApp/QuizApp/Login/EmailView.swift
@@ -1,65 +1,18 @@
 import UIKit
 import SnapKit
 
-class EmailView: UIView {
+class EmailView: TextFieldView {
 
     var delegate: EmailViewDelegate!
-
-    private var textField: UITextField!
 
     override init(frame: CGRect) {
         super.init(frame: frame)
 
-        createViews()
-        styleViews()
-        defineLayoutForViews()
+        setPlaceholder(withText: "Email")
     }
 
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-
-}
-
-extension EmailView: ConstructViewsProtocol {
-
-    func createViews() {
-        textField = UITextField()
-        addSubview(textField)
-    }
-
-    func styleViews() {
-        backgroundColor = .white.withAlphaComponent(0.3)
-        layer.cornerRadius = 20
-        layer.borderColor = UIColor.white.cgColor
-        clipsToBounds = true
-
-        textField.delegate = self
-        textField.attributedPlaceholder = NSAttributedString(
-            string: "Email",
-            attributes: [NSAttributedString.Key.foregroundColor: UIColor.white.withAlphaComponent(0.7)])
-        textField.textColor = .white
-    }
-
-    func defineLayoutForViews() {
-        textField.snp.makeConstraints {
-            $0.leading.equalToSuperview().offset(20)
-            $0.top.equalToSuperview().offset(10)
-            $0.trailing.equalToSuperview().inset(10)
-            $0.bottom.equalToSuperview().inset(10)
-        }
-    }
-
-}
-
-extension EmailView: UITextFieldDelegate {
-
-    func textFieldDidBeginEditing(_ textField: UITextField) {
-        layer.borderWidth = 1
-    }
-
-    func textFieldDidEndEditing(_ textField: UITextField) {
-        layer.borderWidth = 0
     }
 
     func textField(
